Refactor/Reorganize comparison to follow matchers
Breaking apart the ever-growing comparator file into its components. I'm
following the pattern set by Matchers. There's an abstract comparison,
and a bunch of concrete comparisons in their own directory. The concrete
comparisons can decide if they should be built or not. The factory
handles the configuration.
