+++ lib/query_string_search/comparator.rb
@@ -38,57 +38,3 @@ module QueryStringSearch
     end
   end
 end
-
-module QueryStringSearch
-  module Comparator
-    class ComparisonFactory
-      def self.build(config)
-        comparison = if config.subject.respond_to?(:each) || config.other.respond_to?(:each)
-                       SetComparison
-                     elsif [:<, :>, :<=, :>=].include?(config.operator.to_sym)
-                       InequalityComparison
-                     else
-                       EqualityComparison
-                     end
-
-        comparison.new(config.subject, config.other, config.operator.to_sym)
-      end
-    end
-
-    class AbstractComparison
-      attr_accessor :other, :subject, :operator
-
-      def initialize(subject, other, operator = nil)
-        self.subject = subject
-        self.other = other
-        self.operator = operator
-      end
-
-      def normalize(unnormalized)
-        if unnormalized.respond_to?(:each)
-          unnormalized.map(&:to_s).map(&:upcase)
-        else
-          unnormalized.to_s.upcase
-        end
-      end
-    end
-
-    class EqualityComparison < AbstractComparison
-      def compare
-        normalize(other) == normalize(subject)
-      end
-    end
-
-    class SetComparison < AbstractComparison
-      def compare
-        (Array(normalize(subject)) & Array(normalize(other))).any?
-      end
-    end
-
-    class InequalityComparison < AbstractComparison
-      def compare
-        other.to_i.public_send(operator, subject.to_i)
-      end
-    end
-  end
-end
