+++ lib/query_string_search/comparisons/set.rb
@@ -1,0 +1,13 @@
+module QueryStringSearch
+  module Comparator
+    class Set < AbstractComparison
+      def compare
+        (Array(normalize(subject)) & Array(normalize(other))).any?
+      end
+
+      def self.build_me?(config)
+        config.subject.respond_to?(:each) || config.other.respond_to?(:each)
+      end
+    end
+  end
+end
